debug pc sizes

+++ thaumato-anakalyptor/ThaumatoAnakalyptor/sheet_generation/pointcloud_processing.cpp
@@ -984,7 +984,7 @@ private:
             try {
                 std::vector<nf::ResultItem<size_t, double>> ret_matches;
                 nf::SearchParameters params;
-                const double query_pt[3] = { cloud_.pts[i].x, cloud_.pts[i].y, cloud_.pts[i].z };
+                const double query_pt[3] = { static_cast<double>(cloud_.pts[i].x), static_cast<double>(cloud_.pts[i].y), static_cast<double>(cloud_.pts[i].z) };
                 const double radius = spatial_threshold * spatial_threshold;
                 index->radiusSearch(&query_pt[0], radius, ret_matches, params);
                 for (auto& match : ret_matches) {
